Enhanced the networking to be more user-friendly on the first load. Explanation is in the description. Resolves issue #79.
I’ve pushed the CSV fetching into a background thread, after we’ve
checked that there is an available connection to use. So, as a result,
this is what now happens:
 - If there is no connection, show a message and end it there.
 - If there is a connection, check that it’s suitable (up to 10 seconds
wait)
 - Then load the first view correctly, and fetch the data in the
background.

We can now go ahead and setup a loading view with a spinner to show
while this is happening in the background.